# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es mi API",
        "title": "Mi API",
        "contact": {},
        "version": "1"
    },
    "host": "172.25.214.69:60000",
    "basePath": "/Vehicle",
    "paths": {
        "/users/login/": {
            "post": {
                "description": "Autentica a un usuario y devuelve un token de acceso y un token de refresco.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autenticación"
                ],
                "summary": "Iniciar sesión",
                "parameters": [
                    {
                        "description": "Credenciales del usuario",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa con tokens y detalles del usuario",
                        "schema": {
                            "$ref": "#/definitions/views.Response"
                        }
                    },
                    "400": {
                        "description": "Error: Datos inválidos o problema con el formato del email"
                    },
                    "401": {
                        "description": "Error: Credenciales incorrectas o intento de inicio de sesión fallido"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/users/refresh": {
            "post": {
                "description": "Refresca un token de acceso utilizando un token de refresco válido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Refrescar el token de acceso",
                "parameters": [
                    {
                        "description": "Token de refresco",
                        "name": "refreshToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.RefreshTokenForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa con un nuevo token de acceso",
                        "schema": {
                            "$ref": "#/definitions/views.RefreshTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Error: Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Token inválido o expirado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/resources": {
            "get": {
                "description": "Retrieve resources associated with a specific role ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get resources by role ID",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved resources",
                        "schema": {
                            "$ref": "#/definitions/views.ResourceResponse"
                        }
                    },
                    "500": {
                        "description": "Error fetching resources",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "views.LoginForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "views.RefreshTokenForm": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "views.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "accesstoken": {
                    "type": "string"
                },
                "refreshtoken": {
                    "type": "string"
                }
            }
        },
        "views.ResourceResponse": {
            "type": "object",
            "properties": {
                "actions": {},
                "resource": {
                    "type": "string"
                }
            }
        },
        "views.Response": {
            "type": "object",
            "properties": {
                "accesstoken": {
                    "type": "string"
                },
                "failedattempts": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "refreshtoken": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/views.User"
                }
            }
        },
        "views.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fkcompany": {
                    "type": "integer"
                },
                "fkcustomer": {
                    "type": "integer"
                },
                "fkrole": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lastlogin": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}